requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database - Using SQLite for initial setup (MongoDB can be configured later)
# djongo>=1.4.0
# pymongo>=4.0.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0
django-axes==6.1.0

# File Handling
Pillow==10.1.0
python-magic==0.4.27
openpyxl==3.1.2
xlsxwriter==3.1.9

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Utilities
python-decouple==3.8
celery==5.3.4
redis==5.0.1

# Development & Testing
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Documentation
drf-spectacular==0.26.5

# Environment & Configuration
python-dotenv==1.0.0
django-environ==0.11.2
 
# Admin UI Theme
django-jazzmin>=2.6d 